support v0 cid leafs
